+++ CypartaGraphqlSubscriptionsTools/events.py
@@ -1,24 +1,33 @@
+import logging
+
 from channels.layers import get_channel_layer
 from django.core import serializers
+
 from .serialize import serialize_value
+from .utils import GroupNameInvalid, validate_group_name
+
+logger = logging.getLogger(__name__)
+
+
 async def trigger_subscription(group, value):
-    # Print some information for debugging purposes
-    # print('Value:', value)
-    # print('ID:', value.id)
-    # print('Group:', group)
-    
     # Get the channel layer
     channel_layer = get_channel_layer()
+
+    try:
+        group = validate_group_name(group)
+    except GroupNameInvalid:
+        logger.warning(
+            "trigger_subscription skipped: group name failed validation",
+            exc_info=False,
+        )
+        return
 
     # Send a message to the specified channel group
     await channel_layer.group_send(
         group,
         {
             "type": "subscription.triggered",  # Type of the message
-            "value": await serialize_value(value,group),  # The value to be sent
-            "group": group  # The channel group to send the message to
-        }
+            "value": await serialize_value(value, group),  # The value to be sent
+            "group": group,  # The channel group to send the message to
+        },
     )
-
-    # Print a message indicating that the message has been sent
-